+++ src/shared/lib/redux/hooks/use-app-selector.ts
@@ -1,4 +1,4 @@
 import { useSelector, TypedUseSelectorHook } from 'react-redux';
-import { TRootState } from '../types/store-types';
+import { TRootState } from '../../../../types';
 
 export const useAppSelector: TypedUseSelectorHook<TRootState> = useSelector;
