+++ src/shared/lib/redux/selectors/create-status-object.ts
@@ -1,6 +1,6 @@
 import { createSelector } from '@reduxjs/toolkit';
 import { APIStatus } from '../../../enum';
-import { TRootState } from '../types/store-types';
+import { TRootState } from '../../../../types';
 
 export const createStatusObject = (getStatusSelector: (state: TRootState) => APIStatus) => createSelector(
   [getStatusSelector],
