+++ src/shared/lib/redux/thunk/create-app-async-thunk.ts
@@ -1,6 +1,6 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import { AxiosInstance } from 'axios';
-import { TRootState, TAppDispatch } from '../types/store-types';
+import { TRootState, TAppDispatch } from '../../../../types';
 
 export const createAppAsyncThunk = createAsyncThunk.withTypes<{
   state: TRootState;
